+++ lshell/cli.py
@@ -8,6 +8,7 @@ import uuid
 
 from lshell import policy as policy_mode
 from lshell import systemsetup as system_setup
+from lshell import hardeninit as harden_init
 from lshell.checkconfig import CheckConfig
 from lshell.shellcmd import LshellTimeOut, ShellCmd
 
@@ -18,6 +19,8 @@ def main():
         sys.exit(policy_mode.main(sys.argv[2:]))
     if len(sys.argv) > 1 and sys.argv[1] == "setup-system":
         sys.exit(system_setup.main(sys.argv[2:]))
+    if len(sys.argv) > 1 and sys.argv[1] == "harden-init":
+        sys.exit(harden_init.main(sys.argv[2:]))
 
     # Set SHELL and process LSHELL_ARGS env variables.
     os.environ["SHELL"] = os.path.realpath(sys.argv[0])